Moving files to folders to make it easier to read
