+++ src/components/Tool/Appeal/AppealDetails/AppealsMainPanel/AppealsMainPanel.tsx
@@ -1,8 +1,11 @@
 import React from 'react';
 import { TableViewModeEnum } from 'src/components/Shared/Header/ListHeader';
 // import { DynamicContactFlow } from '../ContactFlow/DynamicContactFlow';
-import { AppealsContext, AppealsType } from '../AppealsContext/AppealsContext';
-// import { DynamicContactsList } from '../ContactsList/DynamicContactsList';
+import {
+  AppealsContext,
+  AppealsType,
+} from '../../AppealsContext/AppealsContext';
+import { DynamicContactsList } from '../../List/ContactsList/DynamicContactsList';
 import { AppealsMainPanelHeader } from './AppealsMainPanelHeader';
 
 export const AppealsMainPanel: React.FC = () => {
@@ -21,8 +24,7 @@ export const AppealsMainPanel: React.FC = () => {
       <AppealsMainPanelHeader />
       {!userOptionsLoading &&
         (viewMode === TableViewModeEnum.List ? (
-          // <DynamicContactsList />
-          <h3>List</h3>
+          <DynamicContactsList />
         ) : (
           <h3>Flows</h3>
           // <DynamicContactFlow
