+++ src/components/Tool/Appeal/List/AppealsListFilterPanel/AppealsListFilterPanel.tsx
@@ -14,7 +14,7 @@ import {
   AppealListViewEnum,
   AppealsContext,
   AppealsType,
-} from '../AppealsContext/AppealsContext';
+} from '../../AppealsContext/AppealsContext';
 import { AppealsListFilterPanelButton } from './AppealsListFilterPanelButton';
 import { AppealsListFilterPanelItem } from './AppealsListFilterPanelItem';
